refactor(extensions/podman): begin moving notification code to own class
